% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignorance_uncertainty_regions.R
\name{AdjustedAlphaHelper}
\alias{AdjustedAlphaHelper}
\title{Helper to find the optimal z alpha score for computing uncertainty regions
 for range forecasts - see "Vansteelandt, Stijn, et al. "Ignorance and
 uncertainty regions as inferential tools in a sensitivity analysis."
 Statistica Sinica (2006): 953-979.}
\usage{
AdjustedAlphaHelper(
  c.alpha.star,
  fcst.lower,
  fcst.upper,
  se.lower,
  se.upper,
  alpha
)
}
\arguments{
\item{c.alpha.star:}{candidate z score for uncertainty region}

\item{fcst.lower:}{the lower end of the range forecast}

\item{fcst.upper:}{the upper end of the range forecast}

\item{se.lower:}{the standard error of the upper end range forecast}
}
\value{
The absolute difference between the desired confidence level and
 that for the candidate range z score
}
\description{
Helper to find the optimal z alpha score for computing uncertainty regions
 for range forecasts - see "Vansteelandt, Stijn, et al. "Ignorance and
 uncertainty regions as inferential tools in a sensitivity analysis."
 Statistica Sinica (2006): 953-979.
}
