% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_and_disaggregation.R
\name{AggregationParameters}
\alias{AggregationParameters}
\title{Function to set a list of Disaggregation parameters}
\usage{
AggregationParameters(
  aggregate.to.longest = FALSE,
  periods.agg = NULL,
  agg.fun = "sum",
  disaggregate.format = c(format = "\%a"),
  fit.season.terms = TRUE,
  get.pred.int = TRUE,
  replicates = 600
)
}
\arguments{
\item{aggregate.to.longest:}{whether to aggregate to the longest period
(models that can have <= 1 seasonal period, like arima, are then feasible)}

\item{periods.agg:}{seasonal periods to aggregate over}

\item{fit.season.terms:}{whether to fit terms estimating seasonal effects to
the day of week proportions}

\item{get.pred.int.daily:}{whether to get a prediction interval for predicted
day of week proportions (bootstrap intervals obtained)}

\item{replicates:}{number of bootstrap replicates for above prediction
interval}
}
\value{
: list of parameters for disaggregation to daily
}
\description{
Function to set a list of Disaggregation parameters
}
