% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_or_unknown.R
\name{AnalyzeMissingness}
\alias{AnalyzeMissingness}
\title{Function to analyze missingness mechanism for patterns in the external
regressors, seasonal periods or holidays}
\usage{
AnalyzeMissingness(history, x.reg, alpha.sig)
}
\arguments{
\item{history:}{timeseries history to analyze missingness pattern on}

\item{x.reg:}{external regressors to evaluate missingness pattern on}

\item{alpha.sig:}{the significance level for inferring whether a feature
predicts missingness}
}
\value{
: A model summary indicating whether the missingness depends
  on any of the external regressors
}
\description{
Function to analyze missingness mechanism for patterns in the external
regressors, seasonal periods or holidays
}
