% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_forecast.R
\name{AutoArima}
\alias{AutoArima}
\title{Option for forecasting (aggregated or disaggregated, long or short term)
 Finds the optimal arima forecast}
\usage{
AutoArima(
  history,
  fcst.dts,
  dt.units = "days",
  dt.format = ".AsPOSIXlt",
  periods = c(7),
  periods.agg = NULL,
  periods.trig = c(364),
  pred.level = 0.8,
  transform = "box_cox",
  box.cox.lambda = 1,
  x.reg = NULL,
  x.future = NULL,
  holidays.df = NULL
)
}
\arguments{
\item{x.future}{an optional (but required if x.reg is not null) of future
external regressors}

\item{holidays.df}{a dataframe containing holidays to model as features}

\item{history:}{dataframe containing fields for dt, actual, actual.lower,
actual.upper for forecasting}

\item{fcst.dts:}{named vector containing start and end dt of forecast}

\item{dt.units:}{units for the datetimes (dts) in the history, string}

\item{dt.format:}{string giving a formatting function for the dts}

\item{periods:}{length of seasonal period (in aggregated buckets of
periods.agg)}

\item{periods.agg:}{periods that training data (and forecast will be) are
aggregated over}

\item{periods.trig:}{seasonal periods to fit trigonometric curves to as
external regressors}

\item{pred.level:}{level for the prediction intervals}

\item{transform:}{the transformation applied and thus the inverse to apply
to get forecast on the original scale}

\item{box.cox.lambda:}{the value of the box cox lambda in the event the
transform is 'box_cox'}

\item{x.reg:}{an optional dataframe of external regressors}
}
\value{
: dataframe containing the dt, forecast, forecast lower prediction
  limit and upper prediction limit
}
\description{
Option for forecasting (aggregated or disaggregated, long or short term)
 Finds the optimal arima forecast
}
