% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amortization.R
\name{calibrate_apr}
\alias{calibrate_apr}
\title{Calibrate APR based on a survival curve.}
\usage{
calibrate_apr(S, b = 0.1, n = 12, version = 1, ...)
}
\arguments{
\item{S}{function or numeric atomic vector. Survival curve.}

\item{b}{numeric. Flat forward curve reflecting the cost of funding.}

\item{n}{numeric. Total number of installments (term) in month.}

\item{version}{numeric. Version: 1 - q time; 2 - p time; 3 - number of installment.}
}
\description{
Calibrate APR based on a survival curve.
}

