% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clairvoyant_forecast.R
\name{Clairvoyant}
\alias{Clairvoyant}
\title{Main forecasting function for clairvoyant package}
\usage{
Clairvoyant(
  history,
  training.end.dt = NULL,
  forecast.end.dt = NULL,
  dt.units = "days",
  dt.format = ".AsPOSIXlt",
  imputation.parameters = NULL,
  aggregation.parameters = NULL,
  ensemble.parameters = NULL,
  verbose = TRUE
)
}
\arguments{
\item{history:}{dataframe containing fields for dt_inde and actual to forecast}

\item{training.end.dt:}{last dt to be used in training}

\item{forecast.end.dt:}{dt to forecast until (may be moved back)}

\item{dt.units:}{units for the datetimes (dts) in the history, string}

\item{dt.format:}{string giving a formatting function for the dts}

\item{imputation.parameters:}{list of parameters for imputation obtained
using ImputationParameters()}

\item{disaggregation.parameters:}{list of parameters for daily disaggregation
obtained using DisaggregationParameters()}

\item{ensemble.parameters:}{list of parameters for the ensemble of models
(such as the list of models, external regressors, etc) obtained using
 EnsembleParameters()}

\item{verbose:}{boolean for whether to output error or warning messages}
}
\value{
: list of objects for the training data, steps in
 processing the training data, steps in obtaining the aggregated/unaggreated
 forecast, the final forecast and any available forecast error evaluation
}
\description{
Main forecasting function for clairvoyant package
}
