% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_and_disaggregation.R
\name{DisaggregateForecast}
\alias{DisaggregateForecast}
\title{Disaggregate the weekly total forecast to daily sums using multinomial
 regression}
\usage{
DisaggregateForecast(
  unaggregated.history,
  aggregated.history,
  aggregated.forecast,
  aggregation.parameters = NULL,
  event.parameters = NULL,
  disaggregate.period = 7,
  aggregated.period = 52,
  dt.units = "days",
  dt.format = ".AsPOSIXlt",
  disaggregate.format = c(format = "\%a"),
  pred.level = 0.8
)
}
\arguments{
\item{daily.history:}{dataframe containing fields for dt, actual,
actual.lower actual.upper at the daily level}

\item{aggregated.history:}{dataframe containing the above fields at the
aggregated level}

\item{aggregated.forecast:}{dataframe containing the aggregated forecast (w/
lower + upper prediction limits (dt, forecast, forecast.lower and .upper)}

\item{aggregation.parameters:}{list of aggregation parameters}

\item{event.parameters:}{list of event parameters}

\item{aggregated.period:}{length of longest seasonal period in aggregated
buckets}

\item{pred.level:}{level for the prediction intervals}
}
\value{
: dataframe containing the dt, forecast, forecast lower prediction
  limit and upper prediction limit at the daily level
}
\description{
Disaggregate the weekly total forecast to daily sums using multinomial
 regression
}
