% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{.AsPOSIXlt}
\alias{.AsPOSIXlt}
\title{Internal method for formatting datetimes with the origin
always supplied}
\usage{
.AsPOSIXlt(dt)
}
\value{
datetime object with origin so its compatible with
 numeric datetime objects
}
\description{
Internal method for formatting datetimes with the origin
always supplied
}
