% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_forecast.R
\name{EnsembleParameters}
\alias{EnsembleParameters}
\title{Function to create needed hyperparameters for ensemble forecasting}
\usage{
EnsembleParameters(
  transform = "none",
  periods = c(7),
  periods.trig = c(364),
  models = c("Arima011", "Arima111", "Arima013", "Arima113", "Arima112", "Arima012",
    "AutoArima", "Bsts", "ProphetLinear"),
  x.features = NULL,
  holidays.df = NULL,
  consensus.method = "median",
  range.methods = c("LowerQuartile", "UpperQuartile"),
  pred.level = 0.8
)
}
\arguments{
\item{holidays.df}{a dataframe containing holidays to model as features}

\item{transform:}{the transformation applied to the data to find ensemble
forecast for}

\item{periods:}{length of seasonal periods that are not aggregated over
or fit as trigonometric curves (in periods.trig)}

\item{periods.trig:}{seasonal periods to fit trigonometric curves to as
external regressors}

\item{models:}{vector of the model names included in the ensemble}

\item{x.features:}{matrix of x.features to include as regressors in the
ensemble}

\item{consensus.method:}{the function applied to get the consensus forecast}

\item{pred.level:}{confidence level for prediction/confidence interval}
}
\value{
: A list containing the hyperparameters needed to smooth events and
  estimate their effects
}
\description{
Function to create needed hyperparameters for ensemble forecasting
}
