% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routine.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate the model on training data set and output the performance.}
\usage{
evaluate_model(dat_tr, dat_ts, n_trees = 500, eta = 0.001, max_depth = 6,
  subsample = 0.75, colsample_bytree = 0.75, nthread, output = FALSE,
  model_filename, model_filename_bin, feature_filename, rds_x, rds_y,
  grid_search_filename, dep_var = "Change", n_test_samples = 10)
}
\arguments{
\item{dat_tr}{data.frame.}

\item{dat_ts}{data.frame.}

\item{n_trees}{interger. Number of trees.}

\item{eta}{numeric. Shrinkage parameter.}

\item{max_depth}{integer. Number of maximum depth of tree.}

\item{subsample}{numeric. Percentage of row sampling.}

\item{colsample_bytree}{numeric. Percentage of column sampling.}

\item{nthread}{integer. Number of threads.}

\item{output}{logical. Whether to save the model as texts.}

\item{model_filename}{character. Text file to save the model.}

\item{model_filename_bin}{character. Binary file to save the model.}

\item{feature_filename}{character. Text file to save the features.}

\item{rds_x}{character. RDS file name of test data set (features).}

\item{rds_y}{character. RDS file name of test data set (response).}

\item{grid_search_filename}{character. Output for grid searching.}

\item{dep_var}{character. Name of dependent variable.}

\item{n_test_samples}{integer. Number of samples for unit tests.}
}
\description{
Evaluate the model on training data set and output the performance.
}

