% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_forecast.R
\name{EvaluateClairvoyant}
\alias{EvaluateClairvoyant}
\title{Function to get forecast error for the final daily and weekly forecasts and
 booleans of whether the prediction interval covers the actual value}
\usage{
EvaluateClairvoyant(
  unaggregated.history,
  unaggregated.forecast,
  aggregation.parameters = NULL,
  aggregated.histories = NULL,
  aggregated.forecasts = NULL,
  dt.format = ".AsPOSIXlt",
  longest.period = 364,
  get.unagg.pred.int = TRUE
)
}
\arguments{
\item{daily.history:}{the full daily history (including dates that may be in
the forecast}

\item{daily.forecast:}{the final daily forecast}

\item{weekly.forecast:}{the final weekly forecast}

\item{get.daily.pred.int:}{boolean indicating if there are daily prediction
intervals to be evaluated}
}
\value{
: list of dataframes containing forecast evaluation data
}
\description{
Function to get forecast error for the final daily and weekly forecasts and
 booleans of whether the prediction interval covers the actual value
}
