% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurize_and_transform.R
\name{FeaturizeHolidays}
\alias{FeaturizeHolidays}
\title{Function to obtain holiday features as dummy variables
 to use as external regressors}
\usage{
FeaturizeHolidays(history, holidays.df)
}
\arguments{
\item{history}{dataframe of dts to get holiday features for}

\item{holidays.df}{dataframe of dts and holidays to get features for}
}
\value{
matrix of dummy variables for each holiday to be used as external
  regressors
}
\description{
Function to obtain holiday features as dummy variables
 to use as external regressors
}
