% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_or_unknown.R
\name{FillAnalyzeImpute}
\alias{FillAnalyzeImpute}
\title{Function to impute missing or user specified events to impute}
\usage{
FillAnalyzeImpute(
  history,
  imputation.parameters,
  dt.units = "days",
  dt.format = ".AsPOSIXlt",
  x.reg = NULL
)
}
\arguments{
\item{history:}{timeseries history to impute missing values for}

\item{imputation.parameters:}{an ImputationParameters object giving the
parameters used in imputation}

\item{x.reg:}{the matrix of external regressors to be used in imputing
missing values}
}
\value{
: A list containing the filled history, missingness analysis results,
and the history with imputed missing values (if specified to impute)
}
\description{
Function to impute missing or user specified events to impute
}
