% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_or_unknown.R
\name{FillMissingDts}
\alias{FillMissingDts}
\title{Function to take data frame with potentially missing values for dts and
replace them with fill.default}
\usage{
FillMissingDts(
  history,
  fill.default = NA,
  dt.units = "days",
  dt.format = ".AsPOSIXlt"
)
}
\arguments{
\item{history}{A data frame with two mandatory fields of dt and actual}

\item{fill.default}{To replace the missing values if fill.default is not
provided NA's will be used.}

\item{dt.units}{datetime units in the history}

\item{dt.format}{format function for the datetimes}
}
\value{
The timeseries with missing dates filled by fill.default.
}
\description{
Function to take data frame with potentially missing values for dts and
replace them with fill.default
}
