% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amortization.R
\name{generate_q_time}
\alias{generate_q_time}
\title{Generate remaining principal percentage vector.}
\usage{
generate_q_time(r, n)
}
\arguments{
\item{r}{numeric. Annualized interest rate in decimal.}

\item{n}{numeric. Total number of installments (term) in month.}
}
\description{
Generate remaining principal percentage vector.
}

