% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignorance_uncertainty_regions.R
\name{GetAdjustedAlpha}
\alias{GetAdjustedAlpha}
\title{Function to find the optimal z alpha score for computing uncertainty regions
 for range forecasts - see "Vansteelandt, Stijn, et al. "Ignorance and
 uncertainty regions as inferential tools in a sensitivity analysis."
 Statistica Sinica (2006): 953-979.}
\usage{
GetAdjustedAlpha(range.fcst.df, alpha)
}
\arguments{
\item{range.fcst.df:}{dataframe or named vector containing the forecast
range limits and the corresponding standard errors}

\item{alpha:}{the nominal type I error for the uncertainty regions}
}
\value{
The adjusted Z score given the length of the range forecast
 and the standard errors of the limits of the range forecast
 @export
}
\description{
Function to find the optimal z alpha score for computing uncertainty regions
 for range forecasts - see "Vansteelandt, Stijn, et al. "Ignorance and
 uncertainty regions as inferential tools in a sensitivity analysis."
 Statistica Sinica (2006): 953-979.
}
