% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_forecast.R
\name{GetConsensusForecast}
\alias{GetConsensusForecast}
\title{Function to get a consensus forecast from an ensemble of forecasts}
\usage{
GetConsensusForecast(forecast.list, consensus.method = "median")
}
\arguments{
\item{forecast.list:}{list of forecast dataframes (need to contain fields
for dt, forecast, forecast.lower, foreast.upper)}

\item{consensus.method:}{function to apply to ensemble of forecasts
(pointwise)}
}
\value{
: a dataframe containing the consensus forecast
}
\description{
Function to get a consensus forecast from an ensemble of forecasts
}
