% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurize_and_transform.R
\name{GetHolidayFeatures}
\alias{GetHolidayFeatures}
\title{Function to obtain holiday features as dummy variables
 to use as external regressors (and aggregate as necessary)}
\usage{
GetHolidayFeatures(
  history.start,
  fcst.dts,
  dt.units = "days",
  dt.format = ".AsPOSIXlt",
  periods.agg,
  holidays.df
)
}
\arguments{
\item{history.start}{history start dt to get holiday features for}

\item{fcst.dts}{list of the start and end dts of the forecast}

\item{periods.agg}{vector of periods that the history is aggregated over}

\item{holidays.df}{dataframe of dts and holidays to get features for}
}
\value{
matrix of dummy variables for each holiday to be used as external
  regressors for both the past and future
}
\description{
Function to obtain holiday features as dummy variables
 to use as external regressors (and aggregate as necessary)
}
