% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurize_and_transform.R
\name{GetTrigSeasonalityFeatures}
\alias{GetTrigSeasonalityFeatures}
\title{Function to obtain trigonometric seasonality features to use as
 external regressors (in place of built in dummy variable seasonality)}
\usage{
GetTrigSeasonalityFeatures(history.len, fcst.len, periods)
}
\arguments{
\item{history.len:}{length of the history that the trig seasonality features
are for}

\item{fcst.len:}{length of the forecast that the trig seasonality features
are for}

\item{periods:}{the vector of seasonal periods lengths to create
trigonometric seasonality features of}
}
\value{
matrix of trigonometric seasonality features for each
 seasonal period to be used as external regressors
}
\description{
Function to obtain trigonometric seasonality features to use as
 external regressors (in place of built in dummy variable seasonality)
}
