% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_or_unknown.R
\name{ImputationParameters}
\alias{ImputationParameters}
\title{Function to create needed hyper parameters for any imputation of
missing or other datetimes with unknown values}
\usage{
ImputationParameters(
  missing.fill = NA,
  analyze.missing = TRUE,
  impute.missing = FALSE,
  events.impute = NULL,
  method = "AutoArima",
  transform = "none",
  periods = c(7),
  pred.level = 0.8,
  length.history = 10 * max(periods),
  use.later.history = TRUE
)
}
\arguments{
\item{missing.fill:}{dataframe containing columns for dt and event}

\item{analyze.missing:}{number of days before the event to smooth/estimate effect}

\item{impute.missing:}{number of days after the event to smooth/estimate effect}

\item{events.impute:}{number of weeks before the event to use in smoothing/
effect estimation}

\item{method:}{which forecast function to use to impute if
missing datetimes or other events where the value of the ts is unknown}

\item{transform:}{which}

\item{periods:}{which}

\item{pred.level:}{confidence level for prediction/confidence interval}
}
\value{
: A list containing the hyperparameters needed to smooth events and
  estimate their effects
}
\description{
Function to create needed hyper parameters for any imputation of
missing or other datetimes with unknown values
}
