% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_or_unknown.R
\name{ImputeMissingOrUnknown}
\alias{ImputeMissingOrUnknown}
\title{Function to impute missing or user specified events to impute}
\usage{
ImputeMissingOrUnknown(
  history,
  imputation.parameters,
  dt.format = ".AsPOSIXlt",
  x.reg = NULL
)
}
\arguments{
\item{history:}{timeseries history to impute missing values for}

\item{imputation.parameters:}{an ImputationParameters object giving the
parameters used in imputation}

\item{x.reg:}{the matrix of external regressors to be used in imputing
missing values}
}
\value{
: A list containing the history with imputed missing
  values, and statistics summarizing any user specified events to impute
}
\description{
Function to impute missing or user specified events to impute
}
