% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routine.R
\name{load_data}
\alias{load_data}
\title{Prepare data for financial analysis.}
\usage{
load_data(dir_data, sep, horizon, subsample_freq = 10, ticksize = 0.2,
  cap = Inf, trainpct = 0.9, seed = 29, verbose = TRUE, debug = FALSE)
}
\arguments{
\item{dir_data}{character. The directory of the csv data files.}

\item{sep}{character. The separator of csv file.}

\item{horizon}{integer. The prediction horizon in seconds.}

\item{subsample_freq}{integer. The daily sparsity of subsample frequency.}

\item{ticksize}{numeric. The tick size of the predicted instrument.}

\item{cap}{numeric. Cap in terms of standard deviation.}

\item{trainpct}{numeric. The percentage of training data set.}

\item{seed}{integer. Seed of randomized validation data set.}

\item{verbose}{logical. As name suggests.}

\item{debug}{logical. Whether datetime will be appended to the output of data frame.}
}
\description{
Prepare data for financial analysis.
}

