% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amortization.R
\name{p_to_q}
\alias{p_to_q}
\title{Convert physical installment number to paid principal percentage}
\usage{
p_to_q(i, r, n)
}
\arguments{
\item{i}{numeric. Index of monthly installment.}

\item{r}{numeric. Annualized interest rate in decimal.}

\item{n}{numeric. Total number of installments (term) in month.}
}
\description{
Convert physical installment number to paid principal percentage
}

