% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_forecast.R
\name{PlotErrorComparison}
\alias{PlotErrorComparison}
\title{Function to plot error for two different forecast specifications}
\usage{
PlotErrorComparison(
  object1,
  object2 = NULL,
  label1 = "",
  label2 = "",
  aggregated = FALSE,
  freq = 7,
  dt.format = "as.POSIXct",
  short.term = T,
  n.error = 14,
  log.scale = T,
  dts.annotations = NULL
)
}
\arguments{
\item{object1:}{first clairvoyant forecast object to compare}

\item{object2:}{second clairvoyant forecast object to compare}

\item{label1:}{label for the first forecast specification}

\item{label2:}{label for the 2nd forecast specification}

\item{freq:}{the frequency of the error to plot (daily or weekly)}

\item{short.term:}{whether to truncate to the first n.error measurements}

\item{n.error:}{number of measurements to plot if truncating}

\item{log.scale:}{boolean of whether to plot the error on the log scale}

\item{dts.annotations:}{dataframe of events to annotate the plot}
}
\value{
: null, plot is produced
}
\description{
Function to plot error for two different forecast specifications
}
