% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_forecast.R
\name{PlotForecastEnsemble}
\alias{PlotForecastEnsemble}
\title{Plot the resulting ensemble of models and the consensus forecast}
\usage{
PlotForecastEnsemble(object, dt.format = "as.POSIXct")
}
\arguments{
\item{object:}{clairvoyant forecast object to plot ensemble}
}
\value{
: NULL< plot producted
}
\description{
Plot the resulting ensemble of models and the consensus forecast
}
