% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_forecast.R
\name{PlotTrainingActualAndForecast}
\alias{PlotTrainingActualAndForecast}
\title{Function to plot the training, forecast and actual values}
\usage{
PlotTrainingActualAndForecast(
  object,
  aggregated = FALSE,
  freq = 7,
  plot.range = TRUE,
  short.term = FALSE,
  dt.format = "as.POSIXct",
  dt.units = "days",
  n.history = 28,
  n.forecast = 14
)
}
\arguments{
\item{object:}{clairvoyant forecast object with steps in process}

\item{aggregated:}{whether to plot aggregated forecast}

\item{plot.range:}{whether the range forecast should be plotted}

\item{short.term:}{boolean indicating if should be truncated to short term}

\item{n.history:}{number of training dts to include in short.term plot}

\item{n.forecast:}{number of forecast dts to include in short.term plot}

\item{freq:}{frequency of error (weekly aggregate or daily)}

\item{outl.dir:}{output directory}
}
\value{
: none (plot producted)
}
\description{
Function to plot the training, forecast and actual values
}
