% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{UpperQuartile}
\alias{UpperQuartile}
\title{Function to get upper quartile}
\usage{
UpperQuartile(x, na.rm = TRUE)
}
\arguments{
\item{x:}{vector to get lower quartile of}
}
\value{
: lower quartile of x
}
\description{
Function to get upper quartile
}
