% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{write_config}
\alias{write_config}
\title{Write raw model and feature files to separate config files.}
\usage{
write_config(filename_model, filename_features, filename_tree_prefix,
  filename_features_effective)
}
\arguments{
\item{filename_model}{character. The name of the raw model file.}

\item{filename_features}{character. The name of the raw features file.}

\item{filename_tree_prefix}{character. The prefix path of the config files of trees.}

\item{filename_features_effective}{character. THe name of the effective features file.}
}
\description{
Write raw model and feature files to separate config files.
}

