% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{write_config_alpha}
\alias{write_config_alpha}
\title{Write config alpha file}
\usage{
write_config_alpha(fname_features, output_path, instrument_header,
  instrument_mapper)
}
\arguments{
\item{fname_features}{character. The name of the effective features file.}

\item{output_path}{character. The path to output directory.}

\item{instrument_header}{character. Ticker head of isntrument.}

\item{instrument_mapper}{list. The map between instruments.}
}
\description{
Write config alpha file
}

