% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routine.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate the model on training data set and output the performance.}
\usage{
evaluate_model(dat_tr, dat_ts, n_trees = 500, eta = 0.01, max_depth = 7,
  subsample = 0.75, colsample_bytree = 0.75, nthread, output = FALSE,
  output_path, fname_model = "xgboost.model",
  fname_model_bin = "xgboost.model.bin", fname_feature = "xgboost.features",
  fname_rds_x = "tree.input.rds", fname_rds_y = "tree.output.rds",
  fname_grid_search, dep_var = "Change", n_test_samples = 5000)
}
\arguments{
\item{dat_tr}{data.frame.}

\item{dat_ts}{data.frame.}

\item{n_trees}{interger. Number of trees.}

\item{eta}{numeric. Shrinkage parameter.}

\item{max_depth}{integer. Number of maximum depth of tree.}

\item{subsample}{numeric. Percentage of row sampling.}

\item{colsample_bytree}{numeric. Percentage of column sampling.}

\item{nthread}{integer. Number of threads.}

\item{output}{logical. Whether to save the model as texts.}

\item{output_path}{character. Path to output directory.}

\item{fname_model}{character. Text file to save the model.}

\item{fname_model_bin}{character. Binary file to save the model.}

\item{fname_feature}{character. Text file to save the features.}

\item{fname_rds_x}{character. RDS file name of test data set (features).}

\item{fname_rds_y}{character. RDS file name of test data set (response).}

\item{fname_grid_search}{character. Output for grid searching.}

\item{dep_var}{character. Name of dependent variable.}

\item{n_test_samples}{integer. Number of samples for unit test.}
}
\description{
Evaluate the model on training data set and output the performance.
}

